rename service/main to service/app. just makes more sense.
