+++ robot-server/robot_server/service/dependencies.py
@@ -12,7 +12,7 @@ _rpc_server_instance = None
 
 async def get_hardware() -> HardwareAPILike:
     """Hardware dependency"""
-    from .main import app
+    from .app import app
     # todo Amit 2/11/2020. This function should create and return a singleton
     #  hardware interface.
     return app.extra[HARDWARE_APP_KEY]  # type: ignore
