+++ robot-server/tests/conftest.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock
 
 import pytest
 from starlette.testclient import TestClient
-from robot_server.service.main import app
+from robot_server.service.app import app
 from robot_server.service.dependencies import get_hardware
 from opentrons.hardware_control import API, HardwareAPILike
